app_code/requirements.txt
aggdraw
bleak
matplotlib
numpy
Pillow
Pillow-SIMD
psutil
PySDL2
pysdl2-dll
sounddevice
SoundFile
pyusb
zarr
scipy
opencv-python